feat(cart): Update quantity type to double and add BuyerLoadingSmall widget
- Change quantity parameter type from int to double in CartApiService and related components.
- Introduce BuyerLoadingSmall widget for compact loading indicators in various screens.
- Enhance ingredient list item display by formatting quantity correctly.
- Update payment and product detail handling to accommodate new quantity type.

+++ lib/core/services/cart_api_service.dart
@@ -14,7 +14,7 @@ class CartApiService {
   Future<AddToCartResponse> addToCart({
     required String maNguyenLieu,
     required String maGianHang,
-    required int soLuong,
+    required double soLuong,
     String maCho = 'C01',
   }) async {
     print('🛒 [CART API] ========== ADD TO CART REQUEST ==========');
